+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ## x.y.z
 
 ### Added
+- Add a `getDomDocument()` method
+  ([#630](https://github.com/jjriv/emogrifier/pull/630))
 - Add a Composer script for PHP CS Fixer 
   ([#607](https://github.com/jjriv/emogrifier/pull/607))
 - Copy matching rules with dynamic pseudo-classes or pseudo-elements in